Time: 22 ms (8.81%), Space: 56.6 MB (31.93%) - LeetHub

+++ 2491-divide-players-into-teams-of-equal-skill/2491-divide-players-into-teams-of-equal-skill.cpp
@@ -3,10 +3,10 @@ public:
     long long dividePlayers(vector<int>& skill) {
         sort(skill.begin(),skill.end());
         int i = 0,j=skill.size()-1;
-        int exp = skill[i]+skill[j];
+        long long exp = skill[i]+skill[j];
         long long ans = 0;
         while(i<j){
-            int sum = skill[i]+skill[j];
+            long long sum = skill[i]+skill[j];
             if(sum==exp){
                 ans = (skill[i]*skill[j])+ans;
                 i++,j--;
